Added an intelligent agent for the seekers.
Each individual seeker plays without co-operation.
Thus, this makes it very greedy.

The heuristic is a distance minimization metric which
uses the min weighted cost of the distance to every
location in the belief state.

This can be further optimized by considering prunning
the belief state once the rounds are past a certain
point.